bookstore_project/requirements.txt
Django==4.2.1
djangorestframework==3.14.0
razorpay==2.0.0
reportlab==4.4.4
requests==2.32.5
Pillow==12.0.0
imagehash==4.3.1
numpy==1.24.3
nltk==3.8.1
scikit-learn==1.3.0
opencv-python==4.8.1.78
tensorflow==2.13.0
pandas==2.0.3
